Pull parts of FirebaseAuth into an abstract class.
